PANDO-21: Move `SmallSet` to `Pando.Repositories.Utils` namespace
