+++ src/Pando/Repositories/PandoRepository.cs
@@ -2,6 +2,7 @@
 using System.Collections.Immutable;
 using Pando.DataSources;
 using Pando.Exceptions;
+using Pando.Repositories.Utils;
 using Pando.Serialization;
 
 namespace Pando.Repositories;
